+++ packages/web-server/src/route.test.ts
@@ -811,7 +811,7 @@ suite(describe => {
 				next();
 			}
 
-			@Controller('/abc', { middleware: [counterMiddleware] })
+			@Controller('/abc', { globalMiddleware: [counterMiddleware] })
 			class TestController {
 				@Get('wat')
 				wat() {}
@@ -898,7 +898,7 @@ suite(describe => {
 				next();
 			}
 
-			@Controller('/abc', { middleware: [counterMiddleware] })
+			@Controller('/abc', { globalMiddleware: [counterMiddleware] })
 			class TestController {
 				@Get('wat')
 				wat() {
@@ -930,7 +930,7 @@ suite(describe => {
 				++counter;
 				next();
 			}
-			@Controller('', { middleware: [counterMiddleware] })
+			@Controller('', { globalMiddleware: [counterMiddleware] })
 			class FeatureController {
 				@Get('wat')
 				get() {
@@ -995,7 +995,7 @@ suite(describe => {
 				next();
 			}
 
-			@Controller('/abc', { middleware: [counterMiddleware] })
+			@Controller('/abc', { globalMiddleware: [counterMiddleware] })
 			class TestController {
 				@Get('/wat')
 				wat() {}
@@ -1026,7 +1026,7 @@ suite(describe => {
 				zoom : number;
 			}
 			
-			@Controller('', { middleware: [ counterMiddleware ]})
+			@Controller('', { globalMiddleware: [ counterMiddleware ]})
 			class ApiController {
 				@Post('/info', { middleware: [ bodyParser.json() ] })
 				getX(@Body() body : MyRequestType) {